Add AXUIElement::is_settable
When adding this I noticed a slight discrepancy between the bindings and
docs. The out param of AXUIElementIsAttributeSettable is documented as a
"DarwinBoolean" which is just a C unsigned char and technically could
have any value, not just 0 or 1.

This is different from the return values of functions like
AXIsProcessTrusted, which are documented to return only TRUE or FALSE.